chore: move powers_of_two to shares pkg
