+++ app/estimate_square_size.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/celestiaorg/celestia-app/pkg/appconsts"
 	"github.com/celestiaorg/celestia-app/pkg/shares"
-	"github.com/celestiaorg/celestia-app/pkg/utils"
 	"github.com/cosmos/cosmos-sdk/client"
 	core "github.com/tendermint/tendermint/proto/tendermint/types"
 	coretypes "github.com/tendermint/tendermint/types"
@@ -109,7 +108,7 @@ func estimateSquareSize(txs []*parsedTx, evd core.EvidenceList) (uint64, int) {
 
 	// calculate the smallest possible square size that could contain all the
 	// messages
-	squareSize := utils.RoundUpPowerOfTwo(int(math.Ceil(math.Sqrt(float64(txShares + evdShares + msgShares)))))
+	squareSize := shares.RoundUpPowerOfTwo(int(math.Ceil(math.Sqrt(float64(txShares + evdShares + msgShares)))))
 
 	// the starting square size should at least be the minimum
 	if squareSize < appconsts.MinSquareSize {
@@ -134,7 +133,7 @@ func estimateSquareSize(txs []*parsedTx, evd core.EvidenceList) (uint64, int) {
 		// try the next largest square size if we can't fit all the txs
 		case !fits:
 			// double the square size
-			squareSize = utils.RoundUpPowerOfTwo(squareSize + 1)
+			squareSize = shares.RoundUpPowerOfTwo(squareSize + 1)
 		}
 	}
 }
