+++ pkg/shares/powers_of_two_test.go
@@ -1,4 +1,4 @@
-package utils
+package shares
 
 import (
 	"testing"
