--- pkg/utils/doc.go
@@ -1,5 +1,0 @@
-// utils provides utilities for use in celestia-app packages. A desideratum of this package
-// is to avoid importing any celestia-app packages to avoid circular
-// dependencies in consuming packages. The utilities in this package aren't
-// intended to be used outside of celestia-app.
-package utils
